+++ maitake/tests/scheduler/alloc.rs
@@ -1,5 +1,5 @@
 use super::*;
-use mycelium_util::sync::Lazy;
+use mycelium_util::sync::{Lazy, spin::Mutex};
 
 #[test]
 fn basically_works() {
@@ -65,4 +65,51 @@ fn many_yields() {
     assert_eq!(tick.completed, TASKS);
     assert_eq!(COMPLETED.load(Ordering::SeqCst), TASKS);
     assert!(!tick.has_remaining);
+}
+
+#[test]
+fn steal_blocked() {
+    static SCHEDULER_1: Lazy<StaticScheduler> = Lazy::new(StaticScheduler::new);
+    static SCHEDULER_2: Lazy<StaticScheduler> = Lazy::new(StaticScheduler::new);
+    static MUTEX: Mutex<()> = Mutex::new(());
+    static READY: AtomicBool = AtomicBool::new(false);
+    static IT_WORKED: AtomicBool = AtomicBool::new(false);
+
+    util::trace_init();
+
+    let guard = MUTEX.lock();
+
+    let thread = std::thread::spawn(|| {
+        SCHEDULER_1.spawn(async {
+            READY.store(true, Ordering::Release);
+
+            // block this thread
+            let _guard = MUTEX.lock();
+        });
+
+        SCHEDULER_1.spawn(async {
+            IT_WORKED.store(true, Ordering::Release);
+        });
+
+        SCHEDULER_1.tick()
+    });
+
+    while !READY.load(Ordering::Acquire) {
+        core::hint::spin_loop();
+    }
+
+    assert!(SCHEDULER_1.current_task().is_some());
+
+    let stolen = SCHEDULER_1.try_steal().unwrap().spawn_n(&SCHEDULER_2.get(), 1);
+    assert_eq!(stolen, 1);
+
+    let tick = SCHEDULER_2.tick();
+    assert!(IT_WORKED.load(Ordering::Acquire));
+    assert_eq!(tick.polled, 1);
+    assert_eq!(tick.completed, 1);
+
+    drop(guard);
+    let tick = thread.join().unwrap();
+    assert_eq!(tick.polled, 1);
+    assert_eq!(tick.completed, 1);
 }
